+++ ServiceNow/Public/Get-ServiceNowChangeRequest.ps1
@@ -1,93 +1,61 @@
 function Get-ServiceNowChangeRequest {
     [OutputType([System.Management.Automation.PSCustomObject])]
-    [CmdletBinding(DefaultParameterSetName, SupportsPaging)]
+    [CmdletBinding(DefaultParameterSetName = 'Session', SupportsPaging)]
     Param(
         # Machine name of the field to order by
-        [Parameter(Mandatory = $false)]
-        [string]$OrderBy = 'opened_at',
+        [Parameter()]
+        [string] $OrderBy = 'opened_at',
 
         # Direction of ordering (Desc/Asc)
-        [Parameter(Mandatory = $false)]
+        [Parameter()]
         [ValidateSet('Desc', 'Asc')]
-        [string]$OrderDirection = 'Desc',
+        [string] $OrderDirection = 'Desc',
 
         # Maximum number of records to return
-        [Parameter(Mandatory = $false)]
-        [int]$Limit,
+        [Parameter()]
+        [int] $Limit,
 
         # Fields to return
-        [Parameter(Mandatory = $false)]
+        [Parameter()]
         [Alias('Fields')]
-        [string[]]$Properties,
+        [string[]] $Properties,
 
         # Hashtable containing machine field names and values returned must match exactly (will be combined with AND)
-        [Parameter(Mandatory = $false)]
-        [hashtable]$MatchExact = @{},
+        [Parameter()]
+        [hashtable] $MatchExact = @{},
 
         # Hashtable containing machine field names and values returned rows must contain (will be combined with AND)
-        [Parameter(Mandatory = $false)]
-        [hashtable]$MatchContains = @{},
+        [Parameter()]
+        [hashtable] $MatchContains = @{},
 
         # Whether or not to show human readable display values instead of machine values
-        [Parameter(Mandatory = $false)]
+        [Parameter()]
         [ValidateSet('true', 'false', 'all')]
-        [string]$DisplayValues = 'true',
+        [string] $DisplayValues = 'true',
 
-        [Parameter(ParameterSetName = 'SpecifyConnectionFields', Mandatory = $true)]
+        [Parameter(ParameterSetName = 'SpecifyConnectionFields', Mandatory)]
         [ValidateNotNullOrEmpty()]
         [Alias('ServiceNowCredential')]
-        [PSCredential]$Credential,
+        [PSCredential] $Credential,
 
-        [Parameter(ParameterSetName = 'SpecifyConnectionFields', Mandatory = $true)]
-        [ValidateScript({Test-ServiceNowURL -Url $_})]
+        [Parameter(ParameterSetName = 'SpecifyConnectionFields', Mandatory)]
+        [ValidateScript( { $_ | Test-ServiceNowURL })]
         [Alias('Url')]
-        [string]$ServiceNowURL,
+        [string] $ServiceNowURL,
 
-        [Parameter(ParameterSetName = 'UseConnectionObject', Mandatory = $true)]
+        [Parameter(ParameterSetName = 'UseConnectionObject', Mandatory)]
         [ValidateNotNullOrEmpty()]
-        [hashtable]$Connection
+        [hashtable] $Connection,
+
+        [Parameter(ParameterSetName = 'Session')]
+        [ValidateNotNullOrEmpty()]
+        [hashtable] $ServiceNowSession = $script:ServiceNowSession
     )
 
-    # Query Splat
-    $newServiceNowQuerySplat = @{
-        OrderBy         = $OrderBy
-        MatchExact      = $MatchExact
-        OrderDirection  = $OrderDirection
-        MatchContains   = $MatchContains
-    }
-    $Query = New-ServiceNowQuery @newServiceNowQuerySplat
+    $result = Get-ServiceNowTableEntry @PSBoundParameters -Table 'change_request'
 
-    # Table Splat
-    $getServiceNowTableSplat = @{
-        Table         = 'change_request'
-        Query         = $Query
-        Fields        = $Properties
-        DisplayValues = $DisplayValues
+    If ( $result -and -not $Properties) {
+        $result | ForEach-Object { $_.PSObject.TypeNames.Insert(0, "ServiceNow.ChangeRequest") }
     }
-
-    # Update the Table Splat if the parameters have values
-    if ($null -ne $PSBoundParameters.Connection) {
-        $getServiceNowTableSplat.Add('Connection', $Connection)
-    }
-    elseif ($null -ne $PSBoundParameters.Credential -and $null -ne $PSBoundParameters.ServiceNowURL) {
-        $getServiceNowTableSplat.Add('Credential', $Credential)
-        $getServiceNowTableSplat.Add('ServiceNowURL', $ServiceNowURL)
-    }
-
-    # Only add the Limit parameter if it was explicitly provided
-    if ($PSBoundParameters.ContainsKey('Limit')) {
-        $getServiceNowTableSplat.Add('Limit', $Limit)
-    }
-
-    # Add all provided paging parameters
-    ($PSCmdlet.PagingParameters | Get-Member -MemberType Property).Name | Foreach-Object {
-        $getServiceNowTableSplat.Add($_, $PSCmdlet.PagingParameters.$_)
-    }
-
-    # Perform query and return each object in the format.ps1xml format
-    $Result = Get-ServiceNowTable @getServiceNowTableSplat
-    If (-not $Properties) {
-        $Result | ForEach-Object{$_.PSObject.TypeNames.Insert(0,"ServiceNow.ChangeRequest")}
-    }
-    $Result
+    $result
 }